Allow building the xdebug extension on ubuntu 16.04
This role pulls a PHP extension from the XDEBUG website and attempts to
compile it. In the Debian OS family it includes the build-essentials
package to pull down make, compilers etc to assist with this task,
however it does not pull down the phpize package; included as part of
the php-dev package, but otherwise not mandated by any of the
dependencies of this task.

This commit includes the php-dev package for the Debian OS family, so
that it can compile successfully.